Organize the new classes
