+++ src/main/java/dev/denaro/dialog/Dialog.java
@@ -4,6 +4,9 @@ import dev.denaro.FriendlyGuidePlugin;
 import dev.denaro.dialog.options.*;
 import dev.denaro.dialog.options.requirements.*;
 import dev.denaro.yaml.*;
+import dev.denaro.yaml.types.YamlArray;
+import dev.denaro.yaml.types.YamlObject;
+import dev.denaro.yaml.types.YamlValue;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -15,7 +18,6 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 
 public abstract class Dialog
 {
