+++ src/main/java/dev/denaro/dialog/options/DialogExploreResponse.java
@@ -1,10 +1,7 @@
 package dev.denaro.dialog.options;
 
 import dev.denaro.dialog.options.requirements.DialogRequirement;
-import dev.denaro.yaml.YamlArray;
-import net.runelite.api.Client;
-import net.runelite.api.Quest;
-import net.runelite.api.QuestState;
+import dev.denaro.yaml.types.YamlArray;
 
 import java.util.List;
