+++ src/main/java/dev/denaro/yaml/parser/YamlTokenType.java
@@ -1,0 +1,3 @@
+package dev.denaro.yaml.parser;
+
+public enum YamlTokenType {whitespace, string, array_index, kvp_splitter, comment, new_line, folded_block, literal_block}
