+++ src/main/java/dev/denaro/yaml/types/YamlArray.java
@@ -1,4 +1,4 @@
-package dev.denaro.yaml;
+package dev.denaro.yaml.types;
 
 import java.util.ArrayList;
 import java.util.Iterator;
