+++ src/main/java/dev/denaro/yaml/types/YamlObject.java
@@ -1,4 +1,4 @@
-package dev.denaro.yaml;
+package dev.denaro.yaml.types;
 
 import java.util.HashMap;
