+++ src/main/java/dev/denaro/yaml/types/YamlValue.java
@@ -1,4 +1,4 @@
-package dev.denaro.yaml;
+package dev.denaro.yaml.types;
 
 public abstract class YamlValue
 {
